Cleanup of HOOMDInitializer and HOOMDDumpWriter. The initializer, especially, is much cleaner and it will now be easy to add fields to the file format.
Old hoomd_xml files will not load with the new Initializer. Files in the repository have been updated, but those outside may have issues.

All tags in the xml files are now lowercase. Also, "N" type values in the node attributes have been removed in favor of auto-detecting the length of the input. The file format is now more consistent with the original documentation.
ticket #47


git-svn-id: file:///Users/joaander/devel/hoomd-git-convert/repos/trunk@944 fa922fa7-2fde-0310-acd8-f43f465a7996