Intelligent r_cut handling added.
The maximum r_cut among all type pairs and all pair forces using the neighbor list is determined at the beginning of each run. The determination is determined via a callback, so the pair force can 
change the calculation (i.e. to add a buffer for particles with a 
different diameter)

This is all tested in the updated hoomd_script_tests

git-svn-id: file:///Users/joaander/devel/hoomd-git-convert/repos/branches/template-pair-potentials@2429 fa922fa7-2fde-0310-acd8-f43f465a7996