Fixed bug in box size change detection in update.box_resize
If update.box_resize were given a Variant that remains stable for some time, the updater would still activate every requested step and essentially do nothing to the particles. This results in performance slowdowns if period=1 and the user assumes that update.box_resize is smart enough to do nothing when given a flat variant.

git-svn-id: file:///Users/joaander/devel/hoomd-git-convert/repos/trunk@3982 fa922fa7-2fde-0310-acd8-f43f465a7996