Fixed a bug where integrate.npt had poor start-up performance, even from equilibrated initial configs
Shift initialization of m_curr_group_T and m_curr_P to the first call made
to integrateStepOne(). Prior, m_curr_P was being initialized in the constructor,
when the forces had not yet been evaluated and therefore the pressure completely
invalid. The resulting over-correction from the integrator would send it so
far out of whack that it would take forever to get back. Hopefully this solves
the last of the annoying start-up troubles with npt in hoomd.

git-svn-id: file:///Users/joaander/devel/hoomd-git-convert/repos/trunk@3985 fa922fa7-2fde-0310-acd8-f43f465a7996