RPM build scripts now work with git
The spec file now accepts version, revision, and refspec. Version and
revision are used to determine the name of the resulting rpm. refspec
is the branch/tag to check out for building. It is implemented this way
because the spec file cannot determine what it's version is until it has
checked out the code - and it cannot check out the code until it knows what
the version is!

Makefile gets a new option TAGVER which sets all three of these
appropriately for a tagged release build.

redhat_build_package.sh is updated to handle nightly builds out of the git
repository.