Added unit test for berendsen thermostat
The unit test is extremely simple. It just includes a 1000 particle ideal
gas which it ramps up to temp with a BD thermostat and then runs with a
Berendsen thermostat for 5000 steps of equilibration and 5000 steps of
averaging. The average temperature is checked against the set point.

Not very complicated, but it should suffice as a suitable automatic unit
test of the thermostat.