Reverted to the simpler and more elegant original implementation of ParticleGroup
The original HOOMD implementation of ParticleGroup was simpler and less complicated.
I switched back to it, with a few minor changes:
- rebuildIndexList() now has a GPU version for perfomance reasons
- the consequence is that the internal is_member flag set and the member_tags are now GPUArrays.

The idea for parallel simulations is that the groups are built from the global particle
data via the getPosition()/getType() etc. methods in ParticleData, which are extended
to do MPI communication if necessary. This should also guarantee that the user sees
an unique view of the (global) particle data if using the python interface.