Switch setBox to setGlobalBoxL
This change prevents classes from inadvertantly overwriting the periodic flags
in the global box. Classes must also be careful to base any wrap() calls
off of a copy of the box (getBox) so that they don't perform periodic
boundary math on a box the user requested to be non-periodic.

Currently, there is no initialization for the box periodic flags. That
can be added in another branch. However, I think that the implementation
will be such that the periodic flags are determined once at initialization
and not altered at all during the run.

refs #207