Reinstated the const swap method
I know that this is potentially dangerous (i.e. a class outside ParticleData replaces the positions with a zero-length array). But I am using it in some of my plugins :) We could add some additional safety measures - for example we could assert that all sizes and dimensions are identical between the swapped data.