Extra safety for applying boundary conditions in CommunicatorGPU
It must be made absolutely sure that after applying the _global_ boundary conditions, a
migrating particle ends up in the _local_ box that received it. Due to limited FP precision,
a condition can arise where a particle gets sent through a global boundary but
its coordinates do not get folded by BoxDim. When using BoxDim for PBC we
have to use a shifted global box to wrap the particle.