Change order of boost::python search
Search for libpython-X.Y first, then python or python3. This enables gentoo users with dual python2 and python3 installs to select the python that is not on the path with:
cmake ../ -DPYTHON_EXECUTABLE=/usr/bin/pythonX.y

It should not cause problems for non-gentoo users as those will always fail the first attempt and fall back on python or python3.