Apply additional error checking on the cell list indices
In MPI simulations, asserts appear to be disabled. If the first tolerance
in box check failes at 1.00001, then with a large enough cell list
a ghost particle might end up with a cell index greater than the max
in a dimension - leading to out of bounds memory writes. This fix ensures
that every single memory write goes into a proper cell.

As with the earlier tolerance check, ghost particles that are outside the
range of the cell list are not added into the list. This is fine because
they are out of range of any interactions driven by the cell list.