Revert commits now on branch autotuner-hacking
The recent commits hacking on autotuner/r-buff integration break intentional behavior of the autotuner control and would prevent the autotuner control from working in any simulation that did not start with tune.r_buff. Such drastic experiments should be done in a separate branch.
Also, we do not have a shred of data that the expensive combination of autotuner + r_buff tuning even accomplishes a single .1% of performance benefit.

I left the commits on the branch autotuner-hacks