Added replicate() command to system_data
This is a nice feature that takes a system data and replicates it a number
of times along specified spatial directions. Bonded groups are handled
in a clever way (due to LAMMPS) by using the image flags to unwrap molecules
before replicating them. All particle properties such as velocities etc.
are replicated. Rigid bodies are not supported.