Remove unneeded force compute
CellList::compute() properly forces a recompute of the cell list after a box change. So this manual force is not needed. Leaving the manual force in will cause unintended side effects - for example if the cell list compute() method is called twice in a time step after a box change. With m_force_compute set to true, this will result in the cell list being computed twice where it only needs to be computed once.