Fixed bug (pair force equals zero for external potential plugins on Fermi)
This was the consequence of textures being static global variables with
different storages in different libraries, i.e. in the main
hoomd.so and in the plugin shared library. Because
I had a function in a header file which wasn't inlined, the wrong
texture reference got picked when that function was called from
the plugin code library