Bugfix: init.read_xml on a nonexistant file would hang in MPI runs
I fixed this in a general way. I installed a new excepthook handler that will catch any python exception. The handler calls MPI_Abort in MPI runs. We already had a hook like this, but it was only active during a run() command, and there are some exceptions that can occur outside that.