Add HOOMD_WALLTIME_STOP env var
Set this env var to a unix epoch time value to force hoomd runs to stop
at that time. Use it with limit_multiple to force stops at specific time
steps. run() uses predictive timing based on the current average TPS to
make sure the run will stop before the given time.

To allow for multi-stage runs, when this env var is set, time limited
runs will throw an exception. The user script needs to wrap the runs in
a try, except block to ensure proper cleanup at the end of the scriipt.

There is no documentation for this yet, that will come later. Also, it
currently throws StopIteration. This is a problem if the run is
happening in a for loop. We need a custom exception type to avoid this
problem.

refs #47