Add num_util and setup numpy
This is the first phase in bringing numpy snapshot capability. First, we need the num_util library, to find the numpy headers and to setup the numpy library within the boost framework. This creates a hard dependancy on numpy. I've decided not to make the numpy dep optional because it is almost universally installed, and trivial to install even on systems like Titan.

refs #2