Template particle data snaphots on Real type
Numpy snapshots will be most useful if they are in the native data type
of choise by the user (float or double) - not in whatever mode hoomd was
built in. This commit is phase 1 in offering snapshots of different
types. It templates SnapshotParticleData on a Real type (converting to
vec3<Real> from ScalarN). And updates all code that uses snapshots to
use SnapshotParticleData<Scalar>. Next up: Offer both float and double
snapshots.

refs #2