Add option.parse_args() command
parse_args() parses the command line arguments and then initializaes the
execution configuration. parse_args("string") parses the given string as
the command line arguments instead of sys.argv.

This allows users to write job scripts that specify their own arguments.
As long as the user puts option.parse_args() immediately after "from
hoomd_scrit import *", it ensures that MPI nrank options are initialized
as early as possible.

For now, the old behavior is default and parse_args() is an opt-in
selection.

refs #54