Fix a bug where m_aabbs is not aligned
This leads to seg faults when trying to access m_aabbs within
NeighborListTree. It was in a std::vector. Rather than trying to make a
complex aligned allocator (or depend on boost >=1.56 for its), I
switched the code to use a GPUVector which uses alligned memory. This
GPUVector is not linked to the execution configuration, so it will never
allocate memory on the GPU.