Enable log.query() to get current system properties
Introduce two mode operations to analyze.log. The first mode is the
previous behavior: when logging to a file, log.query() returns the value
cached at the last time step written to the file.

Activate the new mode by providing the filename as `None`. Now query()
will return the current state of the system. With the filename set to
`None`, a period of 1 is implied as the logger cannot guess when the
user might want quantities. This causes energies and virials to be
computed every step and a tiny performance loss as a result.

refs #93