Detect when SLURM_LOCALID is 0 on all ranks
Fall back to gpu_id = global_rank % num_gpus if no other local rank identification method works.
This isn't the most general solution to the problem, but it is the easiest to implment quickly.
The cross section of users running Intel MPI on non-homogenous clusters is probably pretty small.

fixes #108