Add compile instructions for clusters
Remove compile instructions for specific linux distros. These are
impossible to keep up to date as these distros change at an extremely
rapid pace. The hoomd-users list is a better resource for specific
instructions related to a distro.

Also add numpy dependancy to the README.

I removed the build instructions that encouraged building custom python
installs as this is *strongly* discouraged on Lustre file systems. Boost
is the only library that users might need to build themeselves, and I
included updated boost build instructions and reference which national
clusters have useable boost modules and which do not.

refs #128