Move global state into a class in context.py
The new class: SimulationContext stores the system definition, system,
updaters, forces, etc... that define a simulation. The currently active
context is in context.current. The remaining changes in this commit are
a global search and replace for globals with hoomd_script.context.current.

SimulationContext is a context manager with a simple __enter__()
function, it puts itself in context.current. In this way, any number of
SimulationContexts can be active as long as the user can juggle them and
there is enough memory.

context.initialize() creates the first SimulationContext and makes it
current to support the existing job script model.

refs #114
refs #69