Prepend the build dir to the python path in tests
This allows numpy or other libraries to be only available by PYTHONPATH
and the unit tests can still run.

Note that the environment append is done at cmake run time, not when
ctest is run. The PYTHONPATH environment for the test run line is set by
cmake when it runs, and that is when the current PYTHONPATH env var will
be read.

fixes #153