Update P,S arguments to work as documented.
I setup npt to accept both P and S arguments. P is just a scalar
pressure, while S is the tensor stress. If both are specified, P takes
precedence. Not only does this make it so that job scripts require fewer
modifications, but it potentially leads to less confusion as pressure
and stress often refer to different things in nomenclature.

Also replaced tabs with spaces.