ping jenkins to rerun tests
The somewhat hack solution to getting tests passing is to build with double precision on the M6000. Jenkins has been reconfigured to build double, nompi for the m6000 tests. This is fine, as the default is now double precision, we need more of the tests running in that mode.

The permanent solution is to upgrade the jenkins box to CUDA 8.0 (when it is released) which fixes the compiler bug.