It is no longer an error to have unnormalized orientations in hpmc.
This is a problem in large runs with XML files. Text output can result
in quaternions that are slightly not normalized. In a large run on
Titan, a single bad file can take down a whole job. Since HPMC will
just renormalize the quaternion anyways, this should not be an error.
Change to a warning so that users are made aware of issues in their
input file.