Convert compute callback to Nano::Signals
Here is an example of converting a no-return signal to Nano::Signals.
These Communicator hosted callbacks are tricky in that they all have
varying degrees of tracking whether or not they have been connected, so
they are not connected multiple times in multiple runs. We need to be
careful as we update the other Communicator homed signals to maintain
that tracking.

Note that tracking doesn't appear to be necessary for disconnect.
Looking at the Nano::Signals code, it should ignore a disconnect request
for something that is not connected.

Also note that boost signals had num_slots() to test if there were any
connections. Nano::Signals has empty() which has inverted logic.

Further note: These setCommunicator() calls can be quite nasty to work
with. For some odd reason, they do everything with comm and only set
m_comm at the end. This leads to seg faults if you use m_comm in the
method by accident.

refs #126