Make sphinx builds report warnings as errors
Also, fix all such errors in the documenation build. Combined with a
Jenkins job to test building the sphinx documentation, this should
prevent any further occurances of readthedocs failing to update
properly.

The readthedocs sphinx-build does not treat warnings as errors. And, for
some unknonw reason, sphinx treats import errors in autodoc as only
warnings. A failed autodoc import can cause entire components to be
undocumented on readthe docs.

Sometimes, one can test these locally in the source directory:
```python -c "from hoomd import md, hpmc, dem, metal, deprecated"```

However, this is not the same environment as sphinx-build which enables
autodoc_mock for numpy. Debugging import error that occur in
sphin-build, but not on the command line is a painful process and I
don't remember the exact magic commands to do so at this time.

fixes #198