Enable cusolver in CUDA 8 builds
In CUDA 8.0, cusolver makes calls into libgomp but is not linked to it.
NVIDIA does not consider this a bug. The only way we can work around it
is to manually link to libgomp ourselves.

libgomp is not "findable" by cmake because it does not get installed
into a standard path. So I hacked it in with a direct reference to gomp.
I do not know if this works with conda or on Mac, but there it is.

I tested this on linux with clang 3.7 and gcc 4.9.3. I will be testing in
a conda recipe once this commit is live on bitbucket.

This hack will break in the future when clang decides to ship its own
openmp library. I have no idea what this means for intel compiler
support.