Add test to validate hard disks
This test checks the hard disk equation of state at N=256*256
at density 0.698 using reference data from the MPMC paper
where we beat this state point to death with three different
simulation codes.

I have tested this test on maint, where it fails as expected due to
the boxmc bug. On the validation branch with the merged boxmc fix,
it passes in all build configurations. Run time is a few minutes
on the GPU - a bit longer on the CPU (about 10 minutes). I made
the test as short as I could and still fail on the maint branch.