Add documentation for hoomd.hpmc.util.tune
Per issue #215, the poor documentation for util.tune was
exacerbated by confusing implementation. This patch attempts
to address at least the first issue.

Documenting both the general and specific behavior of
tune is potentially confusing. It may be preferable to
refactor the general behavior into a base class with
separate documentation. However, since the only difference
is in the default definition of an object member data
structure which is overridden by an initialization argument,
it may be preferable to separate out the `tunable_map`
into documented module properties.

My sphinx documentation skills are weak and there may be
syntax conventions or sphinx extensions in hoomd that I
am not aware of. There is a sphinx warning whose source
was not obvious to me, and I couldn't figure out if there
was a special python domain sphinx syntax for lambda
functions or functions as arguments.