Warn users about rare issue with default quantities
Warning users is the best that can be done for now. I think the best
route to a permanent solution is to add a placeholder data chunk to flag
that all values are default to be used on frames > 0. This is
non-trivial work and I defer it to a later date.

If this bug is impacting anyone seriously, I can walk you through the
changes needed to fix it.

refs #232