Add pair Lennard-Jones 12-08
Lennard-Jones 12-08
===================

A special adaptation of the Lennard-Jones potential where the exponent
on the attractive contribution of the potential is set to `8` rather than
the usual (and theoretically founded) value of `6`.

Since only the exponent is changed and not the pre-factor this potential
is not compatible with the `mie potential` and has to be explicitly added
to HOOMD-blue. It's use case is mainly related to some special cases for
modeling mineral surfaces.