Set the active CUDA device when needed
When using other CUDA libraries imported by python, the active GPU may
be changed for the current thread. When this occurs, kernel launches
fail because they attempt to access memory that is not on the currently
set GPU.

NVIDIA documents cudaSetDevice() as low-overhead. Almost every, if not
every hoomd GPU code gets a GPUArray handle before calling code on the
GPU. I added a method to ExecutionConfiguration that sets the correct
GPU to be active and made GPUArray call this whenever an array is
aquired. There may be corner cases where this does not catch all places
that the GPU needs to be set active. Aside from a lengthy and boring
code review, the only way to find this is by using the code and seeing
when errors crop up.

refs #236