Compile user specified code using clang
Assuing clang is on the path, users can now provide C++ code to the code
arg of jit.patch.user. This code is inserted into a body of the function
of the proper type and compiled by clang. The LLVM IR is loaded and
ready to execute in the JIT module after jit.patch.user completes
construction.