Support LLVM 3.8, 3.9, 4.0, and 5.0
This intricate dance is needed due to slight API differences from
LLVM version to version. I tested 3.8, 3.9, and 4.0 on ubuntu 16.04 and
5.0 on Arch.

I do *not* plan to build and maintain unit tests running on all these
versions going forward. I will test the version compatible with CUDA and
the latest version that Arch installs.

refs #287