Disable BUILD_JIT for old cmake versions
Old versions of cmake do not build the llvm JIT module with C++11 which causes stdlib ABI errors.
Include an error message to notify the user. I cannot imagine a case where a user wants JIT and
would also be using cmake 3.0 or older, so I am not fixing the issue.