MPCD compiles
Tests fail because the snapshot API has been broken. I don't have time to solve this right now, but will revisit
when I get back from vacation. The prinicple problem is to get a pybind11 handle() to the current class to pass
into pybind11::array method so that it can return a numpy array that points to existing data.