Always use local rank assignment when running nranks>1
Fix support for device/core affinity on systems such as Summit.

This change removes support for a feature that no cluster schedule has used in a long time:
compute-exclusive device assignment. Instead, the local rank is always used to select the GPU
when running on more than one rank. When running on only one rank, and when ENABLE_MPI=off,
compute-exclusive device selection is still used.