Display correct contributor count
Github may pretend that non-github users do not exist on the public facing page
(https://github.com/glotzerlab/hoomd-blue/graphs/contributors), but the API can
count them correctly when explicitly requested to. This appears to use the
.mailmap even though that is not mentioned in the API documentation:
https://developer.github.com/v3/repos/#list-contributors

With this change, the badge currently display 67 contributors which matches
the output of `git shortlog` after applying the mailmap.