Enable hpmc.field.lattice_field to work on groups
In the original implementation of the external lattice field, the
harmonic restraints were applied to every particle in the system. This
change allows the user to pass a particle group to the lattice_field
such that the restraints are only applied to the particles in that
group.

I also cleaned up a lot fo the code in ExternalFieldLattice.h.