Add isometric quotient to loggable quantities
So you can pass "shape_isoperimetric_quotient" to the quantities
argument of hoomd.analyze.log().
I only added this for convex polyhedra since there is already a nice
mass_properties class for these that give you the faces, which makes it
trivial to calculate the isoperimetric quotient.
I added it to the UpdaterShape class since that's where I'm concerned
with using it, but I have a feeling that this should probably be made
available even when you're not running alchemy.

I've only tested it by comparing the output to the values I get from
directly calculating the IQ with euclid, and I get answers that are the
same to numerical precision.
So I'm pretty sure this is correct but I do need to write some proper
unit tests (could easily compare to shapes with known values).