Fix several problems with lj_spheres.py validation test
Fixes:
* Apply long range tail correction used in reference data
* Use more reliable reference data from NIST
* Match particle number more closely to reference data
* Make particles 0 diameter (previously, the lj particles had hard cores!)
* Use proper test validating differences of two means
* Widen confidence interval to 99%
* Drastically shorten runtime (this is possible because the system equilibrates very quickly)

I tested this manually with several low density and several high density values from the NIST dataset.
All -0 tests pass with both 2.5.1 and on add-random123-master. The higher density tests are better
at finding errors (such as the hard core). The lower density tests run much faster but are not
as good at detecting errors.

The tests with only cluster moves fail. At this density, the entire system is always in the same
cluster, so the cluster moves do not change the energy.