Fix alpha_iq bias in shape updater
The previous version was incorrect, since it was not comparing the
change in the isoperimetric quotient, but rather just the absolute value
of it. This commit fixes that error. I had to make some changes to
ShapeMoves to calculate the isperimetric quotient so that it can be used
in the update step.