Undo changes to ExternalFieldLattice.h and field.py
I had made originally made these changes based off a branch where I was
adding a group argument to the lattice field. I forgot about that change
and made this branch off of that branch. However, that change should be
in a separate PR, so I'm reverting those changes back to the version on
master.