Implement filter.Types
I had to change the structure of the class to allow for an arbitrary
selection of types and for the types to be given in a string. The error
checking is currently done through the ParticleData.getTypeByName
method.

Note: Currently testing must be turned off in the build since I did know
how to navigate these changes otherwise. I don't think this is a problem
since we will be migrating to pytest, but this should be noted.