Add set operation filters (difference, intersection, union)
The current implementation computes both internal filters separately
before using the set methods in std::algorithm. This is due to there not
being a method to check a single member for passing through a filter. In
many cases this is desirable since we have to get handles to global
arrays and iterate through them. Not having to do that every time
prevents standard filters from being slow.