Depletion with auxillary variables on the CPU
Tune the number of trial insertions to boost the acceptance rate of systems
that are otherwise difficult to equilibrate.

- We implement a validated acceptance rule for depletion with multiple
insertion attempts (ntrial), by defining a joint probability in terms of
unbiased estimators of the distribution of interest and the distribution of the
estimator sample, and by including the realization of the random sample into
the configuration as an auxillary variable. Marginalization over the auxillary
variable recovers the desired configurational weights. The average over
multiple insertions is performed geometrically, in accordance with the
multiplicative nature of the auxillary variable weights per particle.

- The realization of each auxillary variables is stored as a single random number
generator seed per particle, in the vel.x particle data field, and is updated
along with the trial moves. Subsequent tests will confirm if keeping track of
the seed values is necessary.

- Add validation tests for isotropic and anisotropic particles, using different
values of ntrial.