Update _MoveSizeTuneDefinition's handling of acceptance rate
Returns None when the first access of y (i.e. acceptance rate occurs).
This is necessary to prevent irrelevant data from effecting the
calculated acceptance rate.

This also fixes a bug where the total acceptance rate since the
beginning of a run is used for y if the total steps when accessing y is
the same as before.