Fix bug in SolverStep
We were short circuiting on tuning the list of tunables in
`SolverStep.solve` since the first tunable that wasn't tuned made `all`
false and stop iterating. We now explicitly create a list first and then
pass the list to `all` preventing short circuiting. A comment mentioning
the reason for this is given.