Refactor SecantSolver to be more robust
SecantSolver can now handle when the same y is returned twice for
different x's. We apply a wait and see approach; we wait to tune for n
number of times to see if why is just slow to change. If no change is
observed the (n + 1)-th time we perturb x in the direction of last
change. This procedure continues ad infinitum if no change is found.