Modify the logging subsystem namespace generation
Now we filter namespaces more complexly to produce namespaces we want.
This involves two types of removals: plain namespace removals such as in
removing simulation in `('simulation', 'Simulation')` to get
`('Simulation',)`  and removing nested namespaces when we expose the
classes in the base namespace such as `('update', 'box_resize',
'BoxResize')` moving to `('update', 'BoxResize')`. The difference
between these is that 'simulation' was directly singled out while in the
`BoxResize` example 'update' was flagged so that any non-class namespace
after it is removed automatically.

Both of these filter are implemented in a nested function in
`hoomd.logging._LoggerQuantity._generate_namespace` which is used to
filter the default namespace.