Fix ``hoomd.data.typeconverter.TypeConverterValue`` constructor
We checked for the class inclusion in
``TypeConverterValue._conversion_func_dict`` instead of looking at the
subclass or instance status of the class/instance passed to the
constructor. This commit fixes this and does some rearrangement (for
efficiency considerations) to the constructor. We know correctly check
for the subclass or is instance status of all values passed.

Also rearranges the _conversion_func_dict to be in probable class
likelihood order to improve speed of iteration. Removes list from
special preprocessing type.