Add custom particle filters API.
Creates an abstract base class, ``hoomd.filter.CustomFilter`` that
creates the interface for all custom particle filters. We compose these
custom particle filters which are functors that have a notation of
equality and hashing to ``_hoomd.ParticleFilterCustom`` which is a
subclass of ``_hoomd.ParticleFilter`` which allows the the custom filter
to utlimately be used in a ``ParticleGroup``. The cache of filters uses
exclusively the Python subclass of ``hoomd.filter.CustomFilter``. This
commit also edits the default validation for particle filters to allow
for custom filters. A test is added to ensure proper working of custom
filters in actual simulations.