Prevent ``hoomd.Simulation.run`` from being called in local snapshot
This prevents the likely to be most common error in using local
snapshots as the device that the data is on may switch mutliple times
during a run. This does not prevent operations that can operate on a
simulation outside of a run from invalidating buffers (for intance if
``BoxResize`` could run on the GPU and a user was accessing the data on
a CPU this could overwrite any changes they made. More crutical is this
prevents a user from accessing data that has been freed as the result of
setting the simulation's state to another snapshot, changeing particle
number, etc.