Use only one GPU per rank by default
On a test system with CUDA 11.3, hipSetDeviceFlags(hipDeviceMapHost)
appears to trigger an undocumented device auto-select and init.
This change uses cudaSetValidDevices so that the auto-selection chooses
the correct GPU.

Before this change, a mpirun -n 2 simulation would initialize both GPU 0
and 1 on rank 1 according to nvidia-smi. After this change, each rank
initializes only one GPU.