Modify domain decomposition for 2D systems
Due to the new convention in v3 where systems are considered 2D if `Lz =
0`, the domain decomposition is broken for 2D systems.
This commit addresses that change.
In 3D, the domain decomposition is chosen as the one that minimizes the
surface area between the domains.
In 2D, we choose the decomposition that minimizes the linear perimeter
between domains.

Resolves #1038