WallPotential inherits from _HOOMDBaseObject
This makes the most sense as inheriting from hoomd.md.external.field
does not make sense conceptually as they both branch from external and
are considered different types of external potentials. We could later
chose to make a base external class, but their is no current utility in
doing so.