Add array_view class
The array_view class serves as a view (like string_view) upon an
existent data buffer of fixed (or treated as fixed) size. This is to
expose objects like fixed size arrays and ManagedArrays (of
non-numerical types) to Python.

The view is safe so long as the data pointed to is guarenteed not to
move during the duration of the view. In Python, this means that
array_view objects should not be exposed to users and even for
developers handles to array_views should be discarded as quickly as
possible. However, normal use will not likely result in any errors as
long as a new handle is obtained for each usage.