Expose MD wall arrays to Python and test ArrayView
Using the ArrayView infrastructure, the MD EvaluatorWalls::field_type
is exposed to Python with its fixed sized arrays. Also, the
functionality of ArrayView and the exposed EvaluatorWalls::field_type
are tested in this PR.

Also, add methods to make param_type compatible with setting and getting
through Python via PotentialExternal.