Implement `take_snapshot` and `restore_snapshot` APIs
Per the discussion in #819, users find it confusing that the
snapshot property access returns copies and does not allow direct
modification. Make these explicit methods so that the intent
of the design is clear as is the performance implications.

Also, prevent the type names from changing when restoring
snapshots. Allowing the names to change would lead to parameters
for one type applying to another.