Make default simulation None not non-existent
This makes all _HOOMDBaseObject have a _simulation attribute by default
and by default be set to `None`. This means that existing `if
self._simulation is None` checks had to be moved to `isinstance` checks.
In addition, I created a `_SimulationPlaceHolder` for `SyncedList`
instances which provides more robust check if an object is attempted to
be added to two lists. No new tests are needed and existing tests pass
locally (without GPU or MPI).