Fix and test Pair nlist handling of add and attach
This ensures that a neighborlist cannot be associated with two
integrators at the same time or two simulations. We now copy the
neighbor list when a neighborlist associated with multiple pair
potentials is associated with a detached/removed pair force, and the
pair force is added to a new integrator. This raises a RuntimeWarning
but succeeds allowing for performant reentry of a pair force to the same
integrator (no duplication of neighbor lists), but allows adding to a
new integrator as desired since the pair force was removed.