Only import _jit if llvm_enabled==True
The _jit import was wrapped inside a try–except block, but the except
code was just `continue`, silencing any errors that occured on import
(e.g., on import _jit). This commit removes that logic and instead only
attempts to import _jit if hoomd was built with llvm_enabled = True, and
should therefore make import errors more visible and informative.