Remove unused variable
Now that were are computing the extent as the distance from the center
of the union to the center of the farthest particle, and not the
distance to the edge of the fartest particle, we do not need this
variable. It was causing an error in the CI build through the
Werror=unused-variable flag.