Rename *_union variables to *_constituent
To increase code clarity, this commit changes all of the variables in
the PatchEnergyJITUnion classes to have a suffix that is either
_isotropic or _constituent.

This commit also changes the naming of the "alpha array" variables to
the more instructive "param_array".

I also changed the constructor of the PatchEnergyJITUnion class to not
take in the size of the param_array, so that it now matches the behavior
of the non-union class.