Apply OBB + comment block fixes
The OBB members are no longer required for the PatchEnergy classes and
its derivative (especially not the base class since the OBB tree is
specific to the union classes), so they are removed in this commit.

This commit also moves the code block describing the IntegratorHPMC
class to where that class is defined. It seems like the PatchEnergy
class was inserted between the IntegratorHPMC class and its docstring
mistakenly.