Handle isotropic vs. constituent members
This commit does several things, all under the umbrella of handling
isotropic versus constituent fields/methods properly.

First, it reworks the r_cut log. Now, both the PatchEnergy and its union
class export a property to python called r_cut. This returns r_cut for
the singlet class and the max of r_cut_isotropic and (r_cut_constituent +
the distance to the farthest constituent particle) for the union class.
So now the union class overrides getRCut to return that max value.

Second, we now properly handle passing in param_array_isotropic and
param_array_union for the union class.

Finally, we rename the factory and evaluators *_consituent instead of
*_union in the union class.