Fix CPPUnionPotential on GPU
This commit makes a lot of changes that are all aimed at enabling the
use of the python CPPUnionPotential objects when running on the GPU.

The first fix was in the function that launches the JIT-compiled kernel:
it had the wrong number of arguments 28 instead of the expected 30, so
that the kernel was attempting to read memory that wasn't allocated when
reading any of the last 2 arguements. The fix was to add the correct
arguments, and that got rid of the segfault that was occuring on the GPU
for th union potential.

We then discovered that the JIT code seemed to not get called on the
GPU. Some printf and proper gdb debugging revealed that the OBBTree was
somehow not getting updated properly, so that in the HMPC kernel the
energy functions were not actually getting called (because the tree
contained zero particles). We went through and changed the set*
methods for the PatchEnergyJITUnion class to use ManagedArray arrays
instead of jit::union_params_t arrays. This seemed to fix the issue of
the tree being empty, so that the energy function actually gets called.

We also modified the buildOBBTree function to take a type_id argument,
so that the tree is only updated for a particular particle type. This
removes the need for the m_updated_types member of the union classes.

I also fixed the access to the param_arrays from the JIT code...they
were still called alpha_union, but we changed that behavior so that they
can be accessed with param_array_constituent.

I have a test script that runs a simple test of a 2-particle system of
sticky hard spheres, and get the expected behavior when running on both
the GPU and the CPU (the 2 particles remain "stuck" to each other when
the square well potential is hundreds of kT in depth). Additionally, the
entire suite of unit tests pass as of this commit.